membersihkan comment dan memperbaiki strukur controller
