+++ src/main/java/com/bliblifuture/controller/SKUController.java
@@ -1,0 +1,40 @@
+package com.bliblifuture.controller;
+
+import com.bliblifuture.model.SKU;
+import com.bliblifuture.request.SKURequest;
+import com.bliblifuture.response.BaseResponse;
+import com.bliblifuture.response.ListResponse;
+import com.bliblifuture.service.SKUService;
+import com.bliblifuture.service.UnknownSKUService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@Controller
+public class SKUController {
+    @Autowired
+    SKUService skuService;
+
+    @RequestMapping(value = "api/SKUs", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE,
+            produces = MediaType.APPLICATION_JSON_VALUE)
+    @ResponseBody
+    public BaseResponse createSKU(@RequestBody SKURequest request) {
+        skuService.createSKU(request);
+        BaseResponse response = new BaseResponse(true,"");
+        return response;
+    }
+
+    @RequestMapping(value = "api/SKUs" , method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE )
+    @ResponseBody
+    public ListResponse<SKU> getAllDataSKU(@RequestParam String id) {
+
+        List<SKU> data = skuService.findSKUByStockOpname(id);
+        ListResponse<SKU> dataresponse2 = new ListResponse<>(true, "", data);
+        return dataresponse2;
+    }
+
+
+}
