+++ src/main/java/com/bliblifuture/model/StockOpname.java
@@ -1,11 +1,6 @@
-    package com.bliblifuture.model;
-
-import org.springframework.data.repository.cdi.Eager;
+package com.bliblifuture.model;
 
 import javax.persistence.*;
-import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Table(name="stockOpname")
@@ -16,11 +11,6 @@ public class StockOpname {
     private String waktuPembuatan;
     private int totalQty;
     private int totalSKU;
-//    @OneToMany(fetch = FetchType.EAGER)
-//    private List<SKU> SKUs = new ArrayList<>();
-
-//    @OneToMany(fetch = FetchType.EAGER)
-//    private List<UnknownSKU> unknownSKUs = new ArrayList<>();
 
     public String getStockOpnameId() {
         return stockOpnameId;
@@ -58,67 +48,8 @@ public class StockOpname {
         this.waktuPembuatan = waktuPembuatan;
     }
 
-//    public List<SKU> getSKUs() {
-//        return SKUs;
-//    }
-//
-//
-//    public void setSKUs(List<SKU> SKUs) {
-//        this.SKUs = SKUs;
-//    }
-//
-//    public void addSKU(SKU sku){
-//        this.SKUs.add(sku);
-//    }
-//
-//    public List<UnknownSKU> getUnknownSKUs() {
-//        return unknownSKUs;
-//    }
-//
-//    public void setUnknownSKUs(List<UnknownSKU> unknownSKUs) {
-//        unknownSKUs = unknownSKUs;
-//    }
-//
-//    public void addUnknownSKU(UnknownSKU unknownSKU){
-//        this.unknownSKUs.add(unknownSKU);
-//    }
-//
-//    public int getTotalQty() {
-//        return totalQty;
-//    }
-
-
     public void setTotalQty(int totalQty) {
         this.totalQty = totalQty;
     }
-
-//    public void countTotalQty(){
-//        int total = 0;
-//        for (SKU sku: SKUs) {
-//            total = total + sku.getSystemQty();
-//        }
-//        this.totalQty = total;
-//    }
-
-//    public int getTotalSKU() {
-//        return totalSKU;
-//    }
-
-//    public void setTotalSKU(int totalSKU ) {
-//        this.totalSKU = totalSKU;
-//    }
-//    public void countTotalSKU() {
-//        totalSKU = SKUs.size();
-
-//    }
-
-//    public void deleteUnknownSKU(){
-//        this.unknownSKUs.clear();
-//    }
-
-
-
-
-
 
 }
