--- src/main/java/com/bliblifuture/repository/CounterRepository.java
@@ -1,13 +1,0 @@
-//package com.bliblifuture.repository;
-//
-//import com.bliblifuture.model.Counter;
-//import org.springframework.data.jpa.repository.JpaRepository;
-//import org.springframework.stereotype.Repository;
-//
-//import java.util.List;
-//
-//@Repository
-//public interface CounterRepository extends JpaRepository<Counter, String> {
-//    Counter findByUsername(String username);
-////    List<Counter> findByStatus(String status);
-//}
