+++ src/main/java/com/bliblifuture/repository/StockOpnameRepository.java
@@ -2,7 +2,6 @@ package com.bliblifuture.repository;
 
 import com.bliblifuture.model.StockOpname;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
