+++ src/main/java/com/bliblifuture/repository/WorkListRepository.java
@@ -5,5 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WorkListRepository extends JpaRepository<WorkList, String> {
 
-
 }
