+++ src/main/java/com/bliblifuture/request/StockOpnameRequest.java
@@ -1,9 +1,5 @@
 package com.bliblifuture.request;
 
-import com.bliblifuture.model.SKU;
-import com.bliblifuture.model.StockOpname;
-import com.bliblifuture.model.UnknownSKU;
-
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
