+++ src/main/java/com/bliblifuture/request/WarehouseRequest.java
@@ -1,7 +1,6 @@
 package com.bliblifuture.request;
 
 import java.io.Serializable;
-import java.util.List;
 
 public class WarehouseRequest implements Serializable{
     private String data;
