+++ src/main/java/com/bliblifuture/response/UnknownSKUresponses.java
@@ -1,6 +1,5 @@
 package com.bliblifuture.response;
 
-import com.bliblifuture.model.SKU;
 import com.bliblifuture.model.UnknownSKU;
 
 import java.util.List;
